requirements.txt
asgiref==3.3.4
catboost==0.26
cycler==0.10.0
Django==3.2.4
djangorestframework==3.12.4
graphviz==0.16
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
